requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.27.0
rich>=13.0.0